pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["agents*", "api*", "db*", "engine*", "execution*", "market*", "reporting*", "services*", "strategies*", "utils*"]
exclude = ["tests*", "ui*", "ui-react*", "frontend*", "profiles*", "docs*", "results*"]

[tool.setuptools]
py-modules = ["main", "config", "settings", "portfolio_loader", "start_ui", "runtime_context"]

[project]
name = "prediction-wallet"
version = "1.0.0"
description = "AI-governed portfolio rebalancing agent with deterministic policy validation and audit trails"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Rayan Barreddine", email = "rayan.barreddine@gmail.com" }]
keywords = ["portfolio", "AI", "finance", "rebalancing", "pydantic-ai", "governance"]
requires-python = ">=3.13"
dependencies = [
    # Core AI / Agent
    "pydantic-ai>=1.0.10",
    "google-genai>=1.31.0",
    "anthropic>=0.61.0",
    # Data
    "yfinance>=0.2.36",
    "pandas>=2.0.0",
    "numpy>=1.26.0",
    "scipy>=1.11.0",
    # Reporting
    "reportlab>=4.0.0",
    # API / UI
    "fastapi>=0.135.2",
    "uvicorn>=0.30.0",
    "opentelemetry-api>=1.39.0",
    "opentelemetry-sdk>=1.39.0",
    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.0",
    "PyYAML>=6.0",
    # Optional PostgreSQL (activated when DATABASE_URL is set)
    "psycopg[binary]>=3.2.0",
    "sqlalchemy>=2.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]
